requirements.txt
fastapi
uvicorn
transformers
datasets
peft
faiss-cpu
python-dotenv
sentence-transformers
faiss-cpu
pip freeze | grep openai >> requirements.txt